Modify hooks to get dynamic token and chain svg file src
