+++ packages/widgets/src/index.ts
@@ -43,5 +43,7 @@ export type { PopoverInfoProps } from './shared/components/ui/PopoverInfo';
 export { UtilizationBar } from './shared/components/ui/UtilizationBar';
 export type { UtilizationBarProps } from './shared/components/ui/UtilizationBar';
 export { PairTokenIcons } from './shared/components/ui/token/PairTokenIcon';
+export { useTokenImage } from './shared/hooks/useTokenImage';
+export { useChainImage } from './shared/hooks/useChainImage';
 export { WidgetContainer } from './shared/components/ui/widget/WidgetContainer';
 export { CardAnimationWrapper } from './shared/animation/Wrappers';
